fix: make typebox checker method private
remove the export from 'validateWithTypebox' because the only client of that method is 'validateAgainstSchema'.

'validateWithTypebox' method is moved to 'validateAgainstSchema' file
